Alterations to work standalone
Removed more of the things that depended on the Lumith framework. Removed caching to files as the usage here is intended to eval the code instead of loading it via a file.

Fixed the broken references to $mod_templates.

Removed the dest tag, as it doesn't belong standalone.

Cleaned up the code a bunch. Fixed the caching logic to properly use id if it is set and md5 otherwise.

Altered to use Parse::XJR instead of XML::Bare.