First working version with example
Template::Nepl now works. Yay.

Tweaked the node value fetching to work properly with Parse::XJR.

Made lang by set when creating an instance, as well as pkg.

Added the ability to set the Perl package that the code is evaluated under.

Fixed the eval within () to eval within {}, when using the *{+[code]}* format.
The () would work in fewer cases.